Поменял Id с long на Int

+++ PlaceAgregator/PlaceAgregator.Entities/Admin.cs
@@ -6,7 +6,7 @@
         public string LastName { get; private set; }
         public string FamilyName { get; private set; }
 
-        public long AccountId { get; private set; }
+        public int AccountId { get; private set; }
         public Account Account { get; private set; }
     }
 }
