+++ PlaceAgregator/PlaceAgregator.Entities/Charge.cs
@@ -12,10 +12,10 @@ namespace PlaceAgregator.Entities
             [Description("руб/час")]
             MoneyHour = 2
         }
-        public long PlaceId { get; private set; }
+        public int PlaceId { get; private set; }
         public Place Place { get; private set; }
 
-        public long? RateId { get; private set; }
+        public int? RateId { get; private set; }
         public Rate? Rate { get; private set; }
         public TimeInterval TimeInterval { get; private set; }
         public float Value { get; private set; }
