+++ PlaceAgregator/PlaceAgregator.Entities/PlacePhoto.cs
@@ -2,8 +2,8 @@
 {
     public class PlacePhoto
     {
-        public long PlaceId { get; private set; }
-        public long PhotoId { get; private set; }
+        public int PlaceId { get; private set; }
+        public int PhotoId { get; private set; }
         public Place Place { get; private set; }
         public Photo Photo { get; private set; }
         public int Position { get; private set; }
