+++ PlaceAgregator/PlaceAgregator.Entities/Rate.cs
@@ -2,7 +2,7 @@
 {
     public class Rate : Entity
     {
-        public long PlaceId { get; private set; }
+        public int PlaceId { get; private set; }
         public Place Place { get; private set; }
 
         public TimeInterval TimeInterval { get; private set; }
